+++ src/components/progress-bar/components/Rail.tsx
@@ -1,57 +1,11 @@
-import { FC, memo } from 'react';
-import { uuid } from 'uuidv4';
+import { FC } from 'react';
 
 import { useVideo } from '../../../hooks';
-import { Highlight } from '../../../types';
-import {
-	getPercentFromDuration,
-	getRailSegments,
-} from '../../../utils/highlights';
 
+import { RailsList } from './RailsList';
 import { RailStyled } from './RailStyled';
 
 interface RailProps {}
-interface RailListProps {
-	highlights: Highlight[];
-	videoDuration: number;
-	getHighlightColorBlended?: (colors: string[]) => string | undefined;
-}
-
-export const RailsList: FC<RailListProps> = memo(
-	({ highlights, videoDuration, getHighlightColorBlended }) => {
-		const railSegments = getRailSegments(highlights, videoDuration);
-		const segments = railSegments.map(({ start, end }) => {
-			const startPoint = getPercentFromDuration(start, videoDuration);
-			const width = getPercentFromDuration(end - start, videoDuration);
-			const intersectedSegments = highlights.filter(
-				highlight => start >= highlight.start && end <= highlight.end,
-			);
-			const startColorSegment = highlights.find(
-				({ start: startTime }) => startTime === start,
-			)?.color;
-			const endColorSegment = highlights.find(
-				({ end: endTime }) => endTime === end,
-			)?.color;
-			const colors = intersectedSegments.map(({ color }) => color);
-
-			const color = colors.length
-				? getHighlightColorBlended?.(colors)
-				: undefined;
-			console.log('rails list rerender');
-			return (
-				<RailStyled
-					key={uuid()}
-					startPoint={startPoint}
-					width={width}
-					color={color}
-					startColorSegment={startColorSegment}
-					endColorSegment={endColorSegment}
-				/>
-			);
-		});
-		return <>{segments}</>;
-	},
-);
 
 export const Rail: FC<RailProps> = () => {
 	const { api, highlights, getHighlightColorBlended } = useVideo();
@@ -63,7 +17,7 @@ export const Rail: FC<RailProps> = () => {
 
 	const videoDuration = api?.getDuration?.() || 0;
 
-	// Create Rail from highlight segments
+	// Create a list of rails from highlight segments
 	return (
 		<RailsList
 			highlights={highlights}
